+++ CHANGELOG.rst
@@ -345,7 +345,7 @@ Added:
 
 0.21.1 (2022-01-12)
 -------------------
-- Parameterized `peak_detection.peak_detector` for minimum prominence and width scaling
+- Parameterized ``peak_detection.peak_detector`` for minimum prominence and width scaling
 - Changed default scaling factors to make peak-finding more sensitive
 
 0.20.2 (2022-01-12)